freelancer/requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
boto3==1.40.20
botocore==1.40.20
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.5
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-countries==7.6.1
django-decouple==2.1
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
jmespath==1.0.1
kombu==5.5.4
mysqlclient==2.2.7
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
PyJWT==2.10.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
redis==6.4.0
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
gunicorn